Deliverable module 1 part 5.
Change the rust code with the hyper crate to handle a new parrot
endpoint. Add a screenshot of the curl output running the code.
The difficulty was not to understand the hyper example and adding a new
endpoint but working around the type conversion between ut8, bytes and
so on; I had to look a the solution.